+++ symtabAPI/src/emitElf.C
@@ -137,7 +137,7 @@ bool emitElf::driver(Symtab *obj, string fName){
     unsigned loadSecTotalSize = 0;
     unsigned NOBITStotalsize = 0;
     unsigned shStrTabSizeInc = 0;
-    unsigned dirtySecsInc = 0;
+    int dirtySecsInc = 0;
     unsigned extraAlignSize = 0;
     unsigned nonLoadableNamesSize = 0;
     
@@ -269,7 +269,7 @@ bool emitElf::driver(Symtab *obj, string fName){
 	    if(scncount>oldEhdr->e_shstrndx)
 	    	newshdr->sh_offset += shStrTabSizeInc;
 	}
-    newshdr->sh_offset += dirtySecsInc + extraAlignSize;
+	newshdr->sh_offset += (int) (dirtySecsInc + extraAlignSize);
     if(BSSExpandFlag && newshdr->sh_addr){
         unsigned newOff = newshdr->sh_offset - (newshdr->sh_offset & (pgSize-1)) + (newshdr->sh_addr & (pgSize-1));
         if(newOff < newshdr->sh_offset)
